requirements.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==21.2.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
fastapi==0.65.1
firebase-admin==5.0.0
flatbuffers==1.12
gast==0.4.0
google-api-core==1.28.0
google-api-python-client==2.5.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
h11==0.12.0
h5py==3.1.0
httplib2==0.19.1
idna==2.10
iniconfig==1.1.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
mediapipe==0.8.4.2
msgpack==1.0.2
nanoid==2.0.0
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.5.2.52
opencv-python==4.5.2.52
opt-einsum==3.3.0
packaging==20.9
pluggy==0.13.1
proto-plus==1.18.1
protobuf==3.17.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
starlette==0.14.2
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
uvicorn==0.13.4
Werkzeug==2.0.1
wrapt==1.12.1
